perf(attachments): make ETag variant-aware and short-circuit 304 before file read
ETag previously ignored thumbnail width, violating RFC 7232 by serving
different content under one identifier. The 304 check also ran after
the core handler had already read file bytes, defeating the purpose
of the check. Move ETag computation and If-None-Match short-circuit
into attachments-core.ts (after permission checks, before any read),
and drop the now-redundant duplicate ETag/304 logic in the outer
attachments.ts wrapper.
